# [WIP]faster histogram sum up, also better accuracy. 