--- a/include/LightGBM/meta.h
+++ b/include/LightGBM/meta.h
@@ -108,7 +108,7 @@ public:
     return static_cast<T *>(pv);
   }
 
-  void deallocate(T * const p, const std::size_t n) const {
+  void deallocate(T * const p, const std::size_t) const {
     _mm_free(p);
   }