--- a/src/boosting/gbdt.cpp
+++ b/src/boosting/gbdt.cpp
@@ -19,9 +19,9 @@ namespace LightGBM {
 
 #ifdef TIMETAG
 std::chrono::duration<double, std::milli> boosting_time;
-std::chrono::duration<double, std::milli> train_score_time;
-std::chrono::duration<double, std::milli> out_of_bag_score_time;
-std::chrono::duration<double, std::milli> valid_score_time;
+std::chrono::duration<double, std::milli> train_floatime;
+std::chrono::duration<double, std::milli> out_of_bag_floatime;
+std::chrono::duration<double, std::milli> valid_floatime;
 std::chrono::duration<double, std::milli> metric_time;
 std::chrono::duration<double, std::milli> bagging_time;
 std::chrono::duration<double, std::milli> sub_gradient_time;
@@ -50,9 +50,9 @@ GBDT::GBDT()
 GBDT::~GBDT() {
   #ifdef TIMETAG
   Log::Info("GBDT::boosting costs %f", boosting_time * 1e-3);
-  Log::Info("GBDT::train_score costs %f", train_score_time * 1e-3);
-  Log::Info("GBDT::out_of_bag_score costs %f", out_of_bag_score_time * 1e-3);
-  Log::Info("GBDT::valid_score costs %f", valid_score_time * 1e-3);
+  Log::Info("GBDT::train_score costs %f", train_floatime * 1e-3);
+  Log::Info("GBDT::out_of_bag_score costs %f", out_of_bag_floatime * 1e-3);
+  Log::Info("GBDT::valid_score costs %f", valid_floatime * 1e-3);
   Log::Info("GBDT::metric costs %f", metric_time * 1e-3);
   Log::Info("GBDT::bagging costs %f", bagging_time * 1e-3);
   Log::Info("GBDT::sub_gradient costs %f", sub_gradient_time * 1e-3);
@@ -114,12 +114,11 @@ void GBDT::ResetTrainingData(const BoostingConfig* config, const Dataset* train_
       }
     }
     num_data_ = train_data->num_data();
+    num_gradients_per_class_ = (num_data_ + 7) / 8 * 8;
     // create buffer for gradients and hessians
-    if (objective_function_ != nullptr) {
-      size_t total_size = static_cast<size_t>(num_data_) * num_tree_per_iteration_;
-      gradients_.resize(total_size);
-      hessians_.resize(total_size);
-    }
+    size_t total_gradient_size = static_cast<size_t>(num_gradients_per_class_) * num_tree_per_iteration_;
+    gradients_.resize(total_gradient_size);
+    hessians_.resize(total_gradient_size);
     // get max feature index
     max_feature_idx_ = train_data->num_total_features() - 1;
     // get label index
@@ -325,11 +324,11 @@ void GBDT::UpdateScoreOutOfBag(const Tree* tree, const int cur_tree_id) {
     train_score_updater_->AddScore(tree, bag_data_indices_.data() + bag_data_cnt_, num_data_ - bag_data_cnt_, cur_tree_id);
   }
   #ifdef TIMETAG
-  out_of_bag_score_time += std::chrono::steady_clock::now() - start_time;
+  out_of_bag_floatime += std::chrono::steady_clock::now() - start_time;
   #endif
 }
 
-bool GBDT::TrainOneIter(const score_t* gradient, const score_t* hessian, bool is_eval) {
+bool GBDT::TrainOneIter(const float* gradient, const float* hessian, bool is_eval) {
   // boosting from average prediction. It doesn't work well for classification, remove it for now.
   if (models_.empty()
       && gbdt_config_->boost_from_average
@@ -359,11 +358,19 @@ bool GBDT::TrainOneIter(const score_t* gradient, const score_t* hessian, bool is
     auto start_time = std::chrono::steady_clock::now();
     #endif
     Boosting();
-    gradient = gradients_.data();
-    hessian = hessians_.data();
     #ifdef TIMETAG
     boosting_time += std::chrono::steady_clock::now() - start_time;
     #endif
+  } else {
+    for (int k = 0; k < num_tree_per_iteration_; ++k) {
+      const int bias1 = k * num_gradients_per_class_;
+      const int bias2 = k * num_data_;
+      #pragma omp parallel for schedule(static)
+      for (int i = 0; i < num_gradients_per_class_; ++i) {
+        gradients_[bias1 + i] = gradient[bias2 + i];
+        hessians_[bias1 + i] = hessian[bias2 + i];
+      }
+    }
   }
   #ifdef TIMETAG
   auto start_time = std::chrono::steady_clock::now();
@@ -377,22 +384,15 @@ bool GBDT::TrainOneIter(const score_t* gradient, const score_t* hessian, bool is
     #ifdef TIMETAG
     start_time = std::chrono::steady_clock::now();
     #endif
-    if (gradients_.empty()) {
-      size_t total_size = static_cast<size_t>(num_data_) * num_tree_per_iteration_;
-      gradients_.resize(total_size);
-      hessians_.resize(total_size);
-    }
     // get sub gradients
     for (int cur_tree_id = 0; cur_tree_id < num_tree_per_iteration_; ++cur_tree_id) {
-      auto bias = cur_tree_id * num_data_;
+      size_t bias = static_cast<size_t>(cur_tree_id)* num_gradients_per_class_;
       // cannot multi-threading here.
       for (int i = 0; i < bag_data_cnt_; ++i) {
-        gradients_[bias + i] = gradient[bias + bag_data_indices_[i]];
-        hessians_[bias + i] = hessian[bias + bag_data_indices_[i]];
+        gradients_[bias + i] = gradients_[bias + bag_data_indices_[i]];
+        hessians_[bias + i] = hessians_[bias + bag_data_indices_[i]];
       }
     }
-    gradient = gradients_.data();
-    hessian = hessians_.data();
     #ifdef TIMETAG
     sub_gradient_time += std::chrono::steady_clock::now() - start_time;
     #endif
@@ -403,9 +403,11 @@ bool GBDT::TrainOneIter(const score_t* gradient, const score_t* hessian, bool is
     start_time = std::chrono::steady_clock::now();
     #endif
     std::unique_ptr<Tree> new_tree(new Tree(2));
+    size_t gbias = static_cast<size_t>(cur_tree_id) * num_gradients_per_class_;
     if (class_need_train_[cur_tree_id]) {
       new_tree.reset(
-        tree_learner_->Train(gradient + cur_tree_id * num_data_, hessian + cur_tree_id * num_data_, is_constant_hessian_));
+        tree_learner_->Train(gradients_.data() + gbias,
+                             hessians_.data() + gbias, is_constant_hessian_));
     }
     #ifdef TIMETAG
     tree_time += std::chrono::steady_clock::now() - start_time;
@@ -502,7 +504,7 @@ void GBDT::UpdateScore(const Tree* tree, const int cur_tree_id) {
     train_score_updater_->AddScore(tree, cur_tree_id);
   }
   #ifdef TIMETAG
-  train_score_time += std::chrono::steady_clock::now() - start_time;
+  train_floatime += std::chrono::steady_clock::now() - start_time;
   #endif
   #ifdef TIMETAG
   start_time = std::chrono::steady_clock::now();
@@ -512,7 +514,7 @@ void GBDT::UpdateScore(const Tree* tree, const int cur_tree_id) {
     score_updater->AddScore(tree, cur_tree_id);
   }
   #ifdef TIMETAG
-  valid_score_time += std::chrono::steady_clock::now() - start_time;
+  valid_floatime += std::chrono::steady_clock::now() - start_time;
   #endif
 }