… .travis.yml
… CMakeLists.txt
… src/boosting/goss.hpp
… src/c_api.cpp
… src/io/dataset.cpp
… src/io/dense_bin.hpp
… src/io/dense_nbits_bin.hpp
… src/io/ordered_sparse_bin.hpp
… src/io/sparse_bin.hpp
… src/objective/binary_objective.hpp
… src/objective/multiclass_objective.hpp
… src/objective/rank_objective.hpp
… src/objective/regression_objective.hpp
… src/treelearner/gpu_tree_learner.cpp
… src/treelearner/gpu_tree_learner.h
… src/treelearner/leaf_splits.hpp
… src/treelearner/serial_tree_learner.cpp
… src/treelearner/serial_tree_learner.h
… src/treelearner/voting_parallel_tree_learner.cpp
… windows/LightGBM.vcxproj
… windows/LightGBM.vcxproj.filters