comparison:

speed (average of run 3 times):

|  Data | master branch | this PR |
|----|  ----| ---- |
| Higgs|238 s |  232s |
| Yahoo LTR |163 s | 140 s|
| MS LTR |227 s | 212s |

Accuracy:

Higgs's AUC:

| Metric      |master branch | This PR|
|----|  ----| ---- | 
| AUC|0.845612 | 0.845696|


ndcg at Yahoo LTR:

| Metric      |master branch | This PR |
|----|  ----| ---- | 
| ndcg@1 |0.733789|0.734298|
| ndcg@3 |0.74039 |0.739005|
| ndcg@5 |0.757809|0.756383|
| ndcg@10|0.797109|0.796678|


ndcg at MS LTR:

| Metric      | master branch |This PR |
|----|  ----| ---- | 
| ndcg@1 |0.523594|0.524338|
| ndcg@3 |0.504672|0.505648|
| ndcg@5 |0.510191|0.510115|
| ndcg@10|0.527902|0.527797|