@gugatr0n1c 2017-04-15
@guolinke can you please provide more anfo about "also better acurracy"?
tnx!