@guolinke 2017-04-15
@gugatr0n1c 
The previous code have some sum-up issues related to float point. 
consider following code:
```
double s = 0.0f;
for(int i = 0; i < n ; ++i) s += a[i];
```
when ```i``` become larger, the ```s``` is far large than ```a[i]```. Thus, the sum-up will be not correct.

[Kahan summation](https://en.wikipedia.org/wiki/Kahan_summation_algorithm) can solve this problem. But it have some additional cost. 

So i use a simple solution here:

```
double s = 0.0f;
int i = 0;
while(i < n){
  double ts = 0.0f;
  for(int j= 0; j < 65536 && i < n; ++j, ++i){
    ts += a[i];
  }
  s += ts;
}
```

```ts``` is the sum-up of 65536 element, so add its value to ```s``` will be more accuracy. 

I will paste the benchmark result soon.