@guolinke 2017-04-15
@huanzhang12 
I change the type of histogram from ```double``` to ```float```. It may needs some changes for the GPU codes as well. And the parameter ```gpu_use_dp``` can be removed now.